refactoring due deps cycles
